修改 order entity vo->request&response controller->修改add

+++ restaurant_order_system/src/main/java/com/course/model/request/MenuRequest.java
@@ -18,7 +18,7 @@ public class MenuRequest {
 
     @NotNull
     @Schema(description = "菜單類別", example = "主食")
-    private Short type;
+    private Integer type;
 
     @NotNull
     @Schema(description = "菜單價格", example = "400")
