+++ restaurant_order_system/src/main/java/com/course/model/response/MenuResponse.java
@@ -22,7 +22,7 @@ public class MenuResponse {
     private String name;
 
     @Schema(description = "菜單類別")
-    private Short type;
+    private Integer type;
 
     @Schema(description = "菜單價格")
     private BigDecimal price;
